31.05.2017 (7)

+++ admin/users_delete.php
@@ -1,5 +1,5 @@
 <?php
-require_once('header.php');
+require_once('header_none.php');
 check_permis($users->is_admin());
 $gridfs = new GridFS();
 $id = isset($_GET['id']) ? $_GET['id'] : '';
@@ -10,7 +10,9 @@ if(!$users->is_admin()) {
 	echo 'Không thể xoá Admin. <a href="users.html">Trở về</a>';
 } else {
 	if($users->delete()){
-		$gridfs->id = $u['hinhanh']; $gridfs->delete();
+		if($u['hinhanh']){
+			$gridfs->id = $u['hinhanh']; $gridfs->delete();
+		}
 		transfers_to('users.html?msg=Xoá thành công!');
 	} else {
 		transfers_to('users.html?msg=Không thể xoá!');
@@ -18,4 +20,4 @@ if(!$users->is_admin()) {
 }
 
 ?>
-<?php require_once('footer.php'); ?>
+
